ace_system/pyproject.toml
[project]
name = "ace-adk"
version = "0.1.0"
description = "ACE-ADK: Agentic Context Engineering Agent Development Kit"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "ACE-ADK Team" }
]
keywords = ["ai", "agents", "adk", "google", "gemini", "ace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "google-adk>=1.16.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.110.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ace-adk"
Documentation = "https://github.com/yourusername/ace-adk#readme"
Repository = "https://github.com/yourusername/ace-adk"

[project.scripts]
ace-adk = "main:main"

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "ruff>=0.14.1",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config.py", "main.py"]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
